update cuda kernel

+++ cuda_rasterizer/forward.cu
@@ -296,7 +296,7 @@ __global__ void preprocessCUDA(int P, int D, int M,
 	float mid = 0.5f * (cov.x + cov.z);
 	float lambda1 = mid + sqrt(max(0.1f, mid * mid - det));
 	float lambda2 = mid - sqrt(max(0.1f, mid * mid - det));
-	float my_radius = ceil(3.f * sqrt(max(lambda1, lambda2)));
+	int my_radius = (int) ceil(3.f * sqrt(max(lambda1, lambda2)));
 	
 
 	uint2 rect_min, rect_max;
